requirements.txt
elasticsearch
Flask
pprint
pyyaml
